New custom razorpay hook added
